INIT FOLDER
this folder intializes data which should be already present in the page before opening it

data.js:here we write dta in form of array insdie samplinglisitng and then export it using module.export and the beside it is written inside brackets so it returns an obj

index.js: this is created to write logic for the data initialied by init.js we import mongoose,listing,data. const Listing = require("../models/listing"); this is to acces the collection and insert data into it?

 build a connection then in adding funcion we deleteall already existing inside by listing.deletemany and add all data using intit data which was reuired from the data.js.then we call the adding() function this adds it