NEW ROUT
first we create a new button inside the index rout inside form which has method get-  You’re just requesting a new page or a new form for creating a listing, so no data is being sent. This is why you use the GET method.

create a new rout in app.js which render new.ejs

in new.ejs create a form which send data in action - this send data in req.body which can be used write all text,cntry,...and submit it

inside input we give name for everything as it act as a key value pair
we can acces the data set by using let {title}=req.body.title if name was title

to simplify the process we give name =title[listing]

CREATE ROUT

this handels the post request sent by new ejs form with help of app.post
all data of listing sent by form is presnt in req.bodty which is sent as new listing to listing

then we save the listings

then rediredt to the index page ie /listings